Makefile: simplify the patch to obsolete the dependence
The file Makefile comes from platform and shouldn't be modified
with lots of changes, which might conflict with future release.

It reverted the file back to official one and do update the change
with a filter-out solution not to update the impacted snippets to
keep the change as small as possible.

Change-Id: If139eb9eaa7e617d99872ee977a0229951643772