rc: remove et1macaddr to make wl work for responding eth port
it's a bit tricky to limit the number of devices for et.ko to
make wl.ko to initialize the right eth ports (eth1 for 2.4G and
eth2 for 5G).

et.ko will try detecting et1macaddr for the 2nd device. once the
nvram value is missed, it won't initialize the device, and wl.ko
will use eth1 and eth2 instead of eth2 and eth3.

Change-Id: Ie0b8b965bdf4a9b648cc06d6efe505fe62e38820